simplify processors
