--- apps/github-receiver/src/core/processors/pull-request-event.processor.ts
@@ -1,32 +1,0 @@
-import { ConnectedAccountProvider, CreateEventInput } from '@codeheroes/common';
-import { PullRequestEvent } from '../interfaces/github.interface';
-import { BaseEventProcessor } from './base-event.processor';
-import { PullRequestEventDetails } from '../interfaces/event-details.interface';
-
-export class PullRequestEventProcessor extends BaseEventProcessor {
-  protected async processEvent(): Promise<CreateEventInput> {
-    const payload = this.webhookEvent.payload as PullRequestEvent;
-
-    return {
-      eventId: this.webhookEvent.eventId,
-      action: this.webhookEvent.action,
-      source: this.webhookEvent.source as ConnectedAccountProvider,
-      processed: false,
-      details: {
-        actor: {
-          id: payload.sender.id.toString(),
-          username: payload.sender.login,
-        },
-        repository: {
-          id: payload.repository.id.toString(),
-          name: payload.repository.name,
-          owner: payload.repository.owner.login,
-        },
-        prNumber: payload.number,
-        title: payload.pull_request.title,
-        state: payload.pull_request.state,
-      } as PullRequestEventDetails,
-      eventTimestamp: new Date(payload.pull_request.updated_at).toISOString(),
-    };
-  }
-}
